No surface tesselation (#149)
Surfaces have been fully removed from the system. There are several reasons for it but it is mostly a matter of code simplification and reducing code paths. 

We still need to figure out a proper manner of doing IO. In particular, whether we should be explicit about maintaining face varying data or whether we should obey the convention of the file format (load all glTF as vertex-varying and all OBJ as face varying no matter the content). Also the shape io code is getting complicated enough that it is probably best to use intermediate representation for loading/saving mesh data (like using directly yocto_shape or make_shape_data).